third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGeometryElement.isPointInStroke-01-expected.txt

This is a testharness.js-based test.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, no arguments.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': 1 argument required, but only 0 present.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, non-finite argument.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, functional test.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'stroke-dasharray'.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'stroke-dashoffset'.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'stroke-miterlimit'.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'stroke-linejoin'.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'stroke-linecap'.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'pathLength'.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'vector-effect'.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInStroke, 'visibility' and 'pointer-events' have no effect.
  Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
Harness: the test ran to completion.
